Added Feel-Good-MobileApp project files.
It has been developed in Python, using Kivy Library for front end, and Python for backend. .json file has been used as the database to store user names and passwords of the registered user. When the user logs in and enters his state of mind, like happy, sad, unloved, the application generates quotes relevant to the user's state of mind.
This app can run on Windows, Linux, Android, MacOS, and iOS devices.